processor/requirements.txt
pika==1.3.2
celery==5.4.0
pdfplumber==0.11.4
langchain-text-splitters==0.2.4
google-genai==1.5.0
qdrant-client==1.9.1
tenacity==8.2.3